refactor: extract heap tracing
